msterial costing

+++ backend/Modules/SupplyChain/Database/Migrations/2023_11_05_050000_create_scm_mrrs_table.php
@@ -33,7 +33,7 @@ return new class extends Migration
             $table->tinyInteger('is_completed')->comment('0, 1')->nullable();
             // $table->bigInteger('acc_cash_requisition_id')->nullable();
             $table->string('purchase_center')->nullable();
-            $table->unsignedBigInteger('total_value')->nullable();
+            $table->double('total_value')->nullable();
             $table->timestamps();
         });
     }
